+++ src/niscope/system_tests/test_system_niscope.py
@@ -37,11 +37,11 @@ def test_vi_string_attribute(session):
 def test_read(session):
     test_voltage = 1.0
     test_record_length = 2000
-    test_channels = '0,1'
+    test_channels = range(2)
     test_num_channels = 2
     session.configure_vertical(test_voltage, niscope.VerticalCoupling.AC)
     session.configure_horizontal_timing(50000000, test_record_length, 50.0, 1, True)
-    wfm, wfm_infos = session[test_channels].read(test_record_length)
+    wfm, wfm_infos = session.channels[test_channels].read(test_record_length)
     assert len(wfm) == test_num_channels * test_record_length
     assert len(wfm_infos) == test_num_channels
 
@@ -49,12 +49,12 @@ def test_read(session):
 def test_fetch(session):
     test_voltage = 1.0
     test_record_length = 2000
-    test_channels = '0,1'
+    test_channels = range(2)
     test_num_channels = 2
     session.configure_vertical(test_voltage, niscope.VerticalCoupling.AC)
     session.configure_horizontal_timing(50000000, test_record_length, 50.0, 1, True)
     with session.initiate():
-        wfm, wfm_infos = session[test_channels].fetch(test_record_length)
+        wfm, wfm_infos = session.channels[test_channels].fetch(test_record_length)
     assert len(wfm) == test_num_channels * test_record_length
     assert len(wfm_infos) == test_num_channels
 
@@ -62,7 +62,7 @@ def test_fetch(session):
 def test_fetch_binary8_into(session):
     test_voltage = 1.0
     test_record_length = 2000
-    test_channels = '0,1'
+    test_channels = range(2)
     test_num_channels = 2
     wfm = numpy.ndarray(test_num_channels * test_record_length, dtype=numpy.int8)
     # Initialize with NaN so we can later verify all samples were overwritten by the driver.
@@ -70,7 +70,7 @@ def test_fetch_binary8_into(session):
     session.configure_vertical(test_voltage, niscope.VerticalCoupling.AC)
     session.configure_horizontal_timing(50000000, test_record_length, 50.0, 1, True)
     with session.initiate():
-        wfm_infos = session[test_channels].fetch_into(timeout=5.0, wfm=wfm)
+        wfm_infos = session.channels[test_channels].fetch_into(wfm=wfm)
     for sample in wfm:
         assert not math.isnan(sample)
     assert len(wfm_infos) == test_num_channels
@@ -79,7 +79,7 @@ def test_fetch_binary8_into(session):
 def test_fetch_binary16_into(session):
     test_voltage = 1.0
     test_record_length = 2000
-    test_channels = '0,1'
+    test_channels = range(2)
     test_num_channels = 2
     wfm = numpy.ndarray(test_num_channels * test_record_length, dtype=numpy.int16)
     # Initialize with NaN so we can later verify all samples were overwritten by the driver.
@@ -87,7 +87,7 @@ def test_fetch_binary16_into(session):
     session.configure_vertical(test_voltage, niscope.VerticalCoupling.AC)
     session.configure_horizontal_timing(50000000, test_record_length, 50.0, 1, True)
     with session.initiate():
-        wfm_infos = session[test_channels].fetch_into(timeout=5.0, wfm=wfm)
+        wfm_infos = session.channels[test_channels].fetch_into(wfm=wfm)
     for sample in wfm:
         assert not math.isnan(sample)
     assert len(wfm_infos) == test_num_channels
@@ -96,7 +96,7 @@ def test_fetch_binary16_into(session):
 def test_fetch_binary32_into(session):
     test_voltage = 1.0
     test_record_length = 2000
-    test_channels = '0,1'
+    test_channels = range(2)
     test_num_channels = 2
     wfm = numpy.ndarray(test_num_channels * test_record_length, dtype=numpy.int32)
     # Initialize with NaN so we can later verify all samples were overwritten by the driver.
@@ -104,7 +104,7 @@ def test_fetch_binary32_into(session):
     session.configure_vertical(test_voltage, niscope.VerticalCoupling.AC)
     session.configure_horizontal_timing(50000000, test_record_length, 50.0, 1, True)
     with session.initiate():
-        wfm_infos = session[test_channels].fetch_into(timeout=5.0, wfm=wfm)
+        wfm_infos = session.channels[test_channels].fetch_into(wfm=wfm)
     for sample in wfm:
         assert not math.isnan(sample)
     assert len(wfm_infos) == test_num_channels
@@ -113,7 +113,7 @@ def test_fetch_binary32_into(session):
 def test_fetch_double_into(session):
     test_voltage = 1.0
     test_record_length = 2000
-    test_channels = '0,1'
+    test_channels = range(2)
     test_num_channels = 2
     wfm = numpy.ndarray(test_num_channels * test_record_length, dtype=numpy.float64)
     # Initialize with NaN so we can later verify all samples were overwritten by the driver.
@@ -121,7 +121,7 @@ def test_fetch_double_into(session):
     session.configure_vertical(test_voltage, niscope.VerticalCoupling.AC)
     session.configure_horizontal_timing(50000000, test_record_length, 50.0, 1, True)
     with session.initiate():
-        wfm_infos = session[test_channels].fetch_into(timeout=5.0, wfm=wfm)
+        wfm_infos = session.channels[test_channels].fetch_into(wfm=wfm)
     for sample in wfm:
         assert not math.isnan(sample)
     assert len(wfm_infos) == test_num_channels
@@ -196,7 +196,7 @@ def test_configure_horizontal_timing(session):
 
 
 def test_fetch_read_measurement(session):
-    active_channel = session['0']
+    active_channel = session.channels['0']
     read_measurement = active_channel.read_measurement(niscope.ScalarMeasurement.FREQUENCY)[0]  # fetching first measurement from returned array
     expected_measurement = 10000
     in_range = abs(read_measurement - expected_measurement) <= max(1e-02 * max(abs(read_measurement), abs(expected_measurement)), 0.0)  # https://stackoverflow.com/questions/5595425/what-is-the-best-way-to-compare-floats-for-almost-equality-in-python