+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 * [Unreleased](#unreleased)
-* [0.5.0](#060---2017-12-20)
+* [0.6.0](#060---2017-12-20)
 * [0.5.0](#050---2017-11-27)
 * [0.4.0](#040---2017-11-07)
 * [0.3.0](#030---2017-10-13)
@@ -14,12 +14,31 @@ All notable changes to this project will be documented in this file.
 * ### ALL
     * #### Added
     * #### Changed
+        * Option string can now be a python dictionary instead of a string. It will be converted as needed (Fix #661)
+            * Key/Value pairs approporiate for desired behavior
+                ``` python
+                session = nidmm.Session('Dev1', False, {'simulate': True, 'driver_setup': {'Model': '4071', 'BoardType': 'PXI'}})
+                ```
+        * Repeated capabilities are handled differently. See [#737](https://github.com/ni/nimi-python/issues/737) for discussion
     * #### Removed
 * ### NI-DMM
     * #### Added
     * #### Changed
         * `nidmm.Session()` no longer takes id_query parameter
+        * The following functions timeout or delay parameter now is required to be a `datetime.timedelta()` object:
+            * `configure_multi_point()`
+            * `configure_trigger()`
+            * `fetch()`
+            * `fetch_multi_point()`
+            * `fetch_waveform()`
+            * `read()`
+            * `read_multi_point()`
+            * `read_waveform()`
+        * The following functions return a `datetime.datetime()` object representing the date and time
+            * `get_cal_date_and_time()`
+        * Metadata updated to NI-DMM 17.5
     * #### Removed
+        * `nidmm.Session()` no longer takes id_query parameter
         * Removed these enums and disconnected them from the associated attribute
             * `DCBias` - `DC_BIAS`
             * `OffsetCompensatedOhms` - `OFFSET_COMP_OHMS`
@@ -31,11 +50,31 @@ All notable changes to this project will be documented in this file.
 * ### NI-Switch
     * #### Added
     * #### Changed
+        * The following functions timeout, delay or holdoff parameters now is required to be a `datetime.timedelta()` object:
+            * `configure_scan_trigger()`
+            * `wait_for_debounce()`
+            * `wait_for_scan_complete()`
     * #### Removed
 * ### NI-DCPower
     * #### Added
+        * `channel` repeated capability - See #737 for discussion
     * #### Changed
         * Metadata updated to NI-DCPower 17.6.1
+    * #### Removed
+* ### NI-FGEN
+    * #### Added
+        * Repeated capablilites - See #737 for discussion:
+            * `channel` repeated capability
+            * `markers` repeated capability
+            * `script_triggers` repeated capability
+    * #### Changed
+        * Metadata updated to NI-DCPower 17.6.1
+        * The following functions timeout parameter now is required to be a `datetime.timedelta()` object:
+            * `fetch_multiple()`
+            * `wait_for_event()`
+        * The following functions return a `datetime.datetime()` object representing the date and time
+            * `get_ext_cal_last_date_and_time()`
+            * `get_self_cal_last_date_and_time()`
     * #### Removed
         * Removed these enums and disconnected them from the associated attribute
             * `CurrentLimitAutorange` - `CURRENT_LIMIT_AUTORANGE`
@@ -45,11 +84,29 @@ All notable changes to this project will be documented in this file.
 * ### NI-FGEN
     * #### Added
     * #### Changed
+        * The following functions timeout parameter now is required to be a `datetime.timedelta()` object:
+            * `adjust_sample_clock_relative_delay()`
+            * `wait_until+done()`
+        * The following functions return a `datetime.datetime()` object representing the date and time
+            * `get_ext_cal_last_date_and_time()`
+            * `get_self_cal_last_date_and_time()`
     * #### Removed
 * ### NI-SCOPE
     * #### Added
+        * Repeated capablilites - See #737 for discussion:
+            * `channel` repeated capability
     * #### Changed
         * `niscope.Session()` no longer takes id_query parameter
+        * The following functions timeout, delay or holdoff parameters now is required to be a `datetime.timedelta()` object:
+            * `configure_trigger_digital()`
+            * `configure_trigger_edge()`
+            * `configure_trigger_hysteresis()`
+            * `configure_trigger_software()`
+            * `configure_trigger_video()`
+            * `configure_trigger_window()`
+            * `fetch()`
+            * `fetch_measurement_stats()`
+            * `read()`
     * #### Removed
         * Removed these enums and disconnected them from the associated attribute
             * `BoolEnableDisable` - `P2P_ENABLED`, `P2P_ADVANCED_ATTRIBUTES_ENABLED`, `P2P_ONBOARD_MEMORY_ENABLED`